% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_moon.R
\name{draw_key_moon}
\alias{draw_key_moon}
\title{Moon key glyph for legends}
\usage{
draw_key_moon(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\description{
Draws the legend key glyphs used in \code{geom_moon}.
}
