% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_moon.R
\name{geom_moon}
\alias{geom_moon}
\title{Moon charts}
\usage{
geom_moon(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
The moon geom is used to create moon charts, which are like pie charts except
that the proportions are shown as crescent or gibbous portions of a circle,
like the lit and unlit portions of the moon. As such, they work best with
only one or two groups.
}
\details{
\code{geom_moon} acts like \code{geom_points} in that mutiple moons can be
plotted on the same panel with x and y in the plot's coordinate system, but
size determined independently of the coordinate system. This behavior also
means that the moons will always be circular even if the coordinate system is
not square.

In order to get a full circle with two complementary sections (a crescent and
a gibbous moon), you need to plot two shapes: one with \code{right = TRUE}
and one with \code{right = FALSE}, with \code{ratio} on the second one equal
to \code{1 - ratio} on the first.
}
\examples{
# The default size range when mapping size to a variable is often too small
# to see the moons clearly
ggplot2::ggplot(
  data.frame(x = 1:5, y = 1, size = 1:5, ratio = 1:5 * 0.2),
  ggplot2::aes(x = x, y = y, size = size, ratio = ratio)
) +
  geom_moon() +
  ggplot2::scale_size_continuous(range = c(1, 20))

# To make full moon charts, you need to call geom_moon() twice, once with
# right = TRUE and once with right = FALSE and ratio equal to 1 - ratio
# from the first one  
ggplot2::ggplot(dmeladh) +
  geom_moon(
    x = 0.5, y = 0.5, fill = "forestgreen", color = "forestgreen",
    ggplot2::aes(ratio = AdhF / 100)
  ) +
  geom_moon(
    x = 0.5, y = 0.5, fill = "gold", color = "gold",
    ggplot2::aes(ratio = AdhS / 100), right = FALSE
  ) +
  ggplot2::facet_wrap(~Locality, ncol = 7)

# Moon charts (and pie charts) are sometimes useful on maps when x and y
# cannot be used as aesthetic dimensions because they are already spatial
# dimensions. Overplotting needs to be considered carefully, however.   
ggplot2::ggplot(
  subset(dmeladh, N > 200),
  ggplot2::aes(Longitude, Latitude)
) +
  geom_moon(ggplot2::aes(ratio = AdhF / 100), fill = "black") +
  geom_moon(ggplot2::aes(ratio = AdhS / 100), right = FALSE) +
  ggplot2::coord_fixed()
}
